runner/requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1
scikit-learn==1.3.0
plotly==5.15.0
kaleido==0.2.1
requests==2.31.0
python-dotenv==1.0.0
supabase==1.0.4
reportlab==4.0.4
Pillow==10.0.0
openai==1.3.0